Move Model to contracts folder
