+++ src/composer.json
@@ -32,7 +32,7 @@
         ],
         "psr-4": {
             "Application\\": "application/",
-            "Application\\Console\\": "application/console/",
+            "Application\\Contracts\\": "application/contracts/",
             "Application\\Middleware\\": "application/middleware/",
             "Application\\Providers\\": "application/providers/",
             "Domains\\": "domains/",
