+++ src/entities/Accounts/Models/AccountLink.php
@@ -2,7 +2,7 @@
 
 namespace Entities\Accounts\Models;
 
-use Domains\Model;
+use Application\Contracts\Model;
 
 class AccountLink extends Model
 {
