+++ src/entities/Players/Models/MinecraftPlayer.php
@@ -2,7 +2,7 @@
 
 namespace Entities\Players\Models;
 
-use Domains\Model;
+use Application\Contracts\Model;
 use Entities\Bans\BannableModelInterface;
 
 class MinecraftPlayer extends Model implements BannableModelInterface
